refactor: merge control interaction registries into one module INT-1177
