+++ scenarios/light-control/light-control.mod.js
@@ -10,7 +10,7 @@ var ScenarioState = require('virtual-device-helpers.mod').ScenarioState;
 var Logger = require('logger.mod').Logger;
 
 var vdHelpers = require('virtual-device-helpers.mod');
-var aTable = require('registry-action-resolvers.mod');
+var registry = require('control-interaction-registry.mod');
 var extractMqttTopics =
   require('scenarios-general-helpers.mod').extractMqttTopics;
 var isControlTypeValid =
@@ -132,14 +132,14 @@ LightControlScenario.prototype.defineControlsWaitConfig = function (cfg) {
 function validateLightDevices(lightDevices) {
   for (var i = 0; i < lightDevices.length; i++) {
     var ctrl = lightDevices[i];
-    if (!aTable.actionsTable[ctrl.behaviorType]) {
+    if (!registry.actionsTable[ctrl.behaviorType]) {
       log.error(
         "Light-control validation error: behavior type '{}' not found",
         ctrl.behaviorType
       );
       return false;
     }
-    var reqCtrlTypes = aTable.actionsTable[ctrl.behaviorType].reqCtrlTypes;
+    var reqCtrlTypes = registry.actionsTable[ctrl.behaviorType].reqCtrlTypes;
     if (!isControlTypeValid(ctrl.mqttTopicName, reqCtrlTypes)) {
       log.error(
         "Light-control validation error: control '{}' is not of valid type for '{}'",
@@ -545,9 +545,9 @@ function controlValueEquals(controlValue, resolvedValue) {
 function precomputeLightDeviceTargets(cfg) {
   for (var i = 0; i < cfg.lightDevices.length; i++) {
     var ctrl = cfg.lightDevices[i];
-    var entry = aTable.actionsTable[ctrl.behaviorType];
-    ctrl.resolvedOnValue = entry.launchResolver(null, ctrl.actionValue);
-    ctrl.resolvedOffValue = entry.resetResolver(
+    var entry = registry.actionsTable[ctrl.behaviorType];
+    ctrl.resolvedOnValue = entry.launchHandler(null, ctrl.actionValue);
+    ctrl.resolvedOffValue = entry.resetHandler(
       null,
       resolveResetValue(ctrl)
     );
