+++ scenarios/thermostat/dev-note-arc42.md
@@ -72,7 +72,7 @@ wb-scenarios.conf         -->  scenario-init-main.js
          |                          |
          |                     thermostat.mod.js
          |                          |
-         |                          ├── table-handling-actions.mod (setEnable/setDisable)
+         |                          ├── control-interaction-registry.mod (setEnable/setDisable)
          |                          ├── pid-engine.mod (PID computation)
          |                          |
          v                     [wb-rules engine]
@@ -90,32 +90,32 @@ wb-scenarios.conf         -->  scenario-init-main.js
 
 1. Гистерезис (hysteresis)
 
-Классический гистерезис с мёртвой зоной:
+    Классический гистерезис с мёртвой зоной:
 
-```
-        OFF зона          Мёртвая зона          ON зона
-  ──────────────────┬───────────────────┬──────────────────
-                    |                   |
-              lowerLimit          upperLimit
-         (target - hysteresis)  (target + hysteresis)
-```
+    ```
+            OFF зона          Мёртвая зона          ON зона
+    ──────────────────┬───────────────────┬──────────────────
+                        |                   |
+                lowerLimit          upperLimit
+            (target - hysteresis)  (target + hysteresis)
+    ```
 
-- При `curTemp >= upperLimit` → выключить нагрев
-- При `curTemp <= lowerLimit` → включить нагрев
-- В мёртвой зоне → сохранить текущее состояние (без изменений)
+    - При `curTemp >= upperLimit` → выключить нагрев
+    - При `curTemp <= lowerLimit` → включить нагрев
+    - В мёртвой зоне → сохранить текущее состояние (без изменений)
 
 2. PID (pid)
 
-Используется пропорционально-интегрально-дифференциальный регулятор, выход которого (0–100%)
-преобразуется в длительность включения актуатора внутри фиксированного периода.
+    Используется пропорционально-интегрально-дифференциальный регулятор, выход которого (0–100%)
+    преобразуется в длительность включения актуатора внутри фиксированного периода.
 
-Все параметры PID-режима сгруппированы в объекте `pidSettings`:
+    Все параметры PID-режима сгруппированы в объекте `pidSettings`:
 
-- deadBand — зона нечувствительности (°C). Внутри неё П-, И- и Д-составляющие умножаются на 0.1 для предотвращения микропереключений.
-- pidCoefficients (Kp, Ki, Kd) — коэффициенты регулятора (вложенный объект).
-- pwmPeriodSec — период цикла (целое число секунд). Должен быть больше minOffTimeSec.
-- pidRecalcCycles — PID пересчитывается не каждый цикл, а раз в N периодов (экономия ресурсов, допускается инерционность тепловых процессов).
-- minOnTimeSec / minOffTimeSec — минимальное время включения и выключения за период. Защищает реле от слишком коротких импульсов (например, при 100% мощности гарантируется пауза minOffTimeSec).
+    - deadBand — зона нечувствительности (°C). Внутри неё П-, И- и Д-составляющие умножаются на 0.1 для предотвращения микропереключений.
+    - pidCoefficients (Kp, Ki, Kd) — коэффициенты регулятора (вложенный объект).
+    - pwmPeriodSec — период цикла (целое число секунд). Должен быть больше minOffTimeSec.
+    - pidRecalcCycles — PID пересчитывается не каждый цикл, а раз в N периодов (экономия ресурсов, допускается инерционность тепловых процессов).
+    - minOnTimeSec / minOffTimeSec — минимальное время включения и выключения за период. Защищает реле от слишком коротких импульсов (например, при 100% мощности гарантируется пауза minOffTimeSec).
 
 Алгоритм цикла (runPwmCycle):
